Move library into src/tdom and demo into src.
