+++ src/local.toml
@@ -1,0 +1,11 @@
+name = "PEP 750"
+
+interpreter = "https://koxudaxi.github.io/pyodide/tstrings/pyodide.mjs"
+
+packages = ["dill"]
+
+[files]
+"./tdom/__init__.py" = "./tdom/__init__.py"
+"./tdom/dom.py" = "./tdom/dom.py"
+"./tdom/parser.py" = "./tdom/parser.py"
+"./tdom/utils.py" = "./tdom/utils.py"
